3.3 Move background writer loop function and async operations to `src/scanning/background_writer.rs`
